requirements.txt
pyyaml>=6.0
markdown>=3.4
jinja2>=3.1
networkx>=3.0
pytest>=7.0